backend/requirements.txt
Django
djangorestframework
django-cors-headers
aixplain
reportlab
python-decouple
requests
celery
redis